Production implementation for DynamoDBExecutor (#62)

+++ dynamodb/src/main/scala/zio/dynamodb/BillingMode.scala
@@ -5,6 +5,6 @@ object BillingMode {
   private[dynamodb] final case class Provisioned(pt: ProvisionedThroughput) extends BillingMode
   case object PayPerRequest                                                 extends BillingMode
 
-  def provisioned(readCapacityUnit: Int, writeCapacityUnit: Int): Provisioned =
+  def provisioned(readCapacityUnit: Long, writeCapacityUnit: Long): Provisioned =
     Provisioned(ProvisionedThroughput(readCapacityUnit, writeCapacityUnit))
 }
